requirements.txt
# Required packages for fraud detection analysis
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
plotly==5.15.0
numpy==1.24.3
python-dateutil==2.8.2
kaleido==0.2.1
